fix(frontend): cut four static chains that anchor heavy deps in hot chunks
Four contained cuts from the bundle attribution work:

- ProfilePicture pulled the hedgehog-mode package (~645 KB) everywhere
  via a static StaticHedgehog import; now React.lazy behind the existing
  conditional render.
- The CH-queries debug panel anchored chart.js plugins and CodeSnippet's
  highlight.js languages in the eager menu/shortcut chunks via
  GlobalShortcuts, HelpMenu, AccountMenu, and Superpowers; the module is
  now a lazy facade and the modal content loads on open.
- maxContextLogic statically imported revenueAnalyticsLogic (and through
  its schema guards the 778 KB generated AJV validators) for two async
  listeners that only need it on demand; now dynamic imports.
- sqlEditorLogic statically imported the Query component graph (and
  through it lib/hog's crypto polyfills) to render one preview inside a
  save-as-insight dialog; now React.lazy.

The Max AI Thread component still anchors the Query graph from the
authed shell (maxThreadLogic imports getToolCallDescriptionAndWidget
from Thread.tsx, whose widgets reach the Query stack) — these cuts are
prerequisites for that follow-up to pay off fully.
